+++ src/modules/pages/PagesView.js
@@ -8,6 +8,8 @@ const calendarIcon = require('../../../assets/images/pages/calendar.png');
 const chatIcon = require('../../../assets/images/pages/chat.png');
 const galleryIcon = require('../../../assets/images/pages/gallery.png');
 const profileIcon = require('../../../assets/images/pages/profile.png');
+const loginIcon = require('../../../assets/images/pages/login.png');
+const blogIcon = require('../../../assets/images/pages/blog.png');
 
 export default function PagesScreen(props) {
   return (
@@ -81,11 +83,25 @@ export default function PagesScreen(props) {
         >
           <Image
             resizeMode="contain"
-            source={profileIcon}
+            source={loginIcon}
             tintColor={colors.primary}
             style={styles.itemImage}
           />
           <Text style={styles.itemText}>Login</Text>
+        </TouchableOpacity>
+      </View>
+      <View style={styles.row}>
+        <TouchableOpacity
+          onPress={() => props.navigation.navigate('Blog')}
+          style={styles.blogItem}
+        >
+          <Image
+            resizeMode="contain"
+            source={blogIcon}
+            tintColor={colors.primary}
+            style={styles.itemImage}
+          />
+          <Text style={styles.itemText}>Blog</Text>
         </TouchableOpacity>
       </View>
     </View>
@@ -105,6 +121,17 @@ const styles = StyleSheet.create({
   },
   item: {
     flex: 1,
+    height: 120,
+    paddingVertical: 20,
+    borderColor: colors.primaryLight,
+    borderWidth: 1,
+    borderRadius: 5,
+    alignItems: 'center',
+    justifyContent: 'space-around',
+    marginHorizontal: 5,
+  },
+  blogItem: {
+    width: '31%',
     height: 120,
     paddingVertical: 20,
     borderColor: colors.primaryLight,